Move the scene builder to its own file.
